Refactor: libcrmcluster: move CPG functions/macros to source file
... and rename to be more descriptive (and get rid of some of the last
remaining references to AIS).
